requirements.txt
streamlit>=1.23.1
langchain>=0.1.13
langchain-community>=0.0.26
chromadb>=0.4.24
pypdf
python-dotenv